requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
requests==2.31.0
python-dotenv==1.0.0
starlette==0.27.0
pydantic==1.10.7
typing-extensions==4.5.0
gunicorn==20.1.0
flask-cors==3.0.10
click==8.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
blinker==1.6.2 